Add ptero.env file
This file is used for the app key, and has it's own volume.

Docker will fail to correctly create the volume if this file does not exist on startup.
Instead of creating the file, docker will create a directory.

Another solution would have been to create a seperate directory for this single file.
I have decided it's easier to simply commit this empty file.